cloud-functions/generate-response/requirements.txt
functions-framework>=3.0.0
flask>=2.0.0
google-cloud-aiplatform>=1.71.0
vertexai>=1.71.0